tech: rename transform
