+++ api/src/llm/infrastructure/streaming/transforms/response-object-to-event-stream-transform.js
